init/ initialize repo with gulp, jasmine, and basic FamilyTree class
Jasmine tests are located in /spec and are used to drive the development of the project. Initially, the tests only cover the basics of the FamilyTree class.

Gulp is used as a task runner; it watches all .js files and continually lints and tests the code.

All main .js files are located in the /src folder